requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
alabaster==0.7.13
alembic==1.11.2
altair==5.0.1
anyio==3.6.2
appnope==0.1.3
APScheduler==3.6.3
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
black==23.7.0
bleach==6.0.0
blinker==1.6.2
boto3==1.28.31
botocore==1.31.31
Brotli==1.0.9
cachetools==4.2.2
certifi==2022.12.7
certipy==0.1.3
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography==41.0.3
curio==1.6
cycler==0.11.0
dacite==1.7.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docopt==0.6.2
docrepr==0.2.0
docutils==0.18.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg==1.4
fiftyone==0.21.6
fiftyone-brain==0.13.1
fiftyone-db==0.4.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
glob2==0.7
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphql-core==3.2.3
grpcio==1.54.2
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hdfs==2.7.2
hide-code==0.7.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.16.4
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.6.0
inflate64==0.3.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.21.3
ipyparallel==8.6.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.4
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterhub==4.0.1
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kaggle==1.5.13
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==2.0.5
mongoengine==0.24.2
more-itertools==10.0.0
motor==3.2.0
mpmath==1.3.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.7.3
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pamela==1.1.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.0
plotly==5.15.0
pluggy==1.2.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py7zr==0.20.6
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydeck==0.8.1b0
Pygments==2.14.0
pymongo==4.4.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyppmd==1.0.0
pyrsistent==0.19.3
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
pyzstd==0.15.9
qtconsole==5.4.3
QtPy==2.3.1
Quandl==3.7.0
rarfile==4.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.19
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.2
starlette==0.31.0
strawberry-graphql==0.138.1
streamlit==1.23.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.8.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
universal-analytics-python3==1.1.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
voxel51-eta==0.11.0
waitress==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
